+++ src/components/footer.js
@@ -1,0 +1,38 @@
+import React from "react";
+import { Link } from "react-router-dom";
+import Obfuscate from "./obfuscate.js";
+import { github, discord } from "../consts";
+import { useTranslation } from 'react-i18next';
+
+function Footer() {
+    const { t } = useTranslation("home");
+
+    return (
+        <div className="footer">
+        <div>
+          <Link className="footersides link" to="/privacy.html">
+            <Obfuscate>{t("privacy")}</Obfuscate>
+          </Link>
+          <Link className="footersides link" to="/credits.html">
+            <Obfuscate>{t("credits")}</Obfuscate>
+          </Link>
+        </div>
+        <div className="footermiddle">
+          <Obfuscate>© Metallic 2022 - </Obfuscate>{new Date().getFullYear()}
+        </div>
+        <div>
+          <a className="footersides link" href={discord}>
+            <Obfuscate>Discord</Obfuscate>
+          </a>
+          <a
+            className="footersides link"
+            href={github}
+          >
+            <Obfuscate>Github</Obfuscate>
+          </a>
+        </div>
+      </div>
+    )
+}
+
+export default Footer;